flyttet resten av løse funksjoner til fetcher.
