+++ cmd/fetch/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"log/slog"
@@ -12,12 +11,6 @@ import (
 
 	"github.com/jonmartinstorm/reposnusern/internal/fetcher"
 )
-
-type TreeFile struct {
-	Path string `json:"path"`
-	URL  string `json:"url"`
-	Type string `json:"type"`
-}
 
 func main() {
 	logger := slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{
@@ -98,17 +91,17 @@ func main() {
 		ciConfig := []map[string]string{}
 
 		tree := fetcher.GetJSONMap(fmt.Sprintf("https://api.github.com/repos/%s/git/trees/%s?recursive=1", fullName, r["default_branch"].(string)), token)
-		treeFiles := parseTree(tree)
+		treeFiles := fetcher.ParseTree(tree)
 
 		for _, tf := range treeFiles {
 			lpath := strings.ToLower(tf.Path)
 			switch {
-			case isDependencyFile(lpath):
-				appendFile(result["files"].(map[string][]map[string]string), path.Base(tf.Path), tf, fullName, token)
+			case fetcher.IsDependencyFile(lpath):
+				fetcher.AppendFile(result["files"].(map[string][]map[string]string), path.Base(tf.Path), tf, fullName, token)
 			case strings.HasPrefix(path.Base(lpath), "dockerfile"):
-				appendFile(result["files"].(map[string][]map[string]string), path.Base(tf.Path), tf, fullName, token)
+				fetcher.AppendFile(result["files"].(map[string][]map[string]string), path.Base(tf.Path), tf, fullName, token)
 			case strings.HasPrefix(tf.Path, ".github/workflows/"):
-				appendCI(&ciConfig, tf, fullName, token)
+				fetcher.AppendCI(&ciConfig, tf, fullName, token)
 			case tf.Path == "SECURITY.md":
 				result["security"].(map[string]bool)["has_security_md"] = true
 			case tf.Path == ".github/dependabot.yml":
@@ -131,67 +124,4 @@ func main() {
 	allBytes, _ := json.MarshalIndent(allData, "", "  ")
 	_ = os.WriteFile(outputFile, allBytes, 0644)
 	slog.Info("Lagret samlet analyse", "file", outputFile)
-}
-
-func getGitBlob(url, token string) string {
-	var result map[string]interface{}
-	if err := fetcher.GetJSONWithRateLimit(url, token, &result); err != nil {
-		return ""
-	}
-	if content, ok := result["content"].(string); ok {
-		d, _ := base64.StdEncoding.DecodeString(strings.ReplaceAll(content, "\n", ""))
-		return string(d)
-	}
-	return ""
-}
-
-func appendFile(files map[string][]map[string]string, key string, tf TreeFile, repo, token string) {
-	content := getGitBlob(tf.URL, token)
-	files[key] = append(files[key], map[string]string{
-		"path":    tf.Path,
-		"content": content,
-	})
-}
-
-func appendCI(ciList *[]map[string]string, tf TreeFile, repo, token string) {
-	content := getGitBlob(tf.URL, token)
-	*ciList = append(*ciList, map[string]string{
-		"path":    tf.Path,
-		"content": content,
-	})
-}
-
-func parseTree(tree map[string]interface{}) []TreeFile {
-	files := []TreeFile{}
-	if tree == nil {
-		return files
-	}
-	if arr, ok := tree["tree"].([]interface{}); ok {
-		for _, item := range arr {
-			entry := item.(map[string]interface{})
-			if entry["type"] == "blob" {
-				files = append(files, TreeFile{
-					Path: entry["path"].(string),
-					URL:  entry["url"].(string),
-					Type: entry["type"].(string),
-				})
-			}
-		}
-	}
-	return files
-}
-
-func isDependencyFile(p string) bool {
-	files := []string{
-		"package.json", "pom.xml", "build.gradle", "build.gradle.kts",
-		"go.mod", "cargo.toml", "requirements.txt", "pyproject.toml",
-		"composer.json", ".csproj", "gemfile", "gemfile.lock",
-		"yarn.lock", "pnpm-lock.yaml", "package-lock.json",
-	}
-	for _, f := range files {
-		if strings.HasSuffix(p, f) {
-			return true
-		}
-	}
-	return false
 }
