+++ internal/fetcher/github.go
@@ -12,6 +12,12 @@ import (
 	"time"
 )
 
+type TreeFile struct {
+	Path string `json:"path"`
+	URL  string `json:"url"`
+	Type string `json:"type"`
+}
+
 // GetJSONWithRateLimit henter JSON fra en URL og respekterer GitHub rate-limiting.
 func GetJSONWithRateLimit(url, token string, out interface{}) error {
 	for {
@@ -66,4 +72,67 @@ func GetReadme(fullName, token string) string {
 		return string(decoded)
 	}
 	return ""
+}
+
+func getGitBlob(url, token string) string {
+	var result map[string]interface{}
+	if err := GetJSONWithRateLimit(url, token, &result); err != nil {
+		return ""
+	}
+	if content, ok := result["content"].(string); ok {
+		d, _ := base64.StdEncoding.DecodeString(strings.ReplaceAll(content, "\n", ""))
+		return string(d)
+	}
+	return ""
+}
+
+func AppendFile(files map[string][]map[string]string, key string, tf TreeFile, repo, token string) {
+	content := getGitBlob(tf.URL, token)
+	files[key] = append(files[key], map[string]string{
+		"path":    tf.Path,
+		"content": content,
+	})
+}
+
+func AppendCI(ciList *[]map[string]string, tf TreeFile, repo, token string) {
+	content := getGitBlob(tf.URL, token)
+	*ciList = append(*ciList, map[string]string{
+		"path":    tf.Path,
+		"content": content,
+	})
+}
+
+func ParseTree(tree map[string]interface{}) []TreeFile {
+	files := []TreeFile{}
+	if tree == nil {
+		return files
+	}
+	if arr, ok := tree["tree"].([]interface{}); ok {
+		for _, item := range arr {
+			entry := item.(map[string]interface{})
+			if entry["type"] == "blob" {
+				files = append(files, TreeFile{
+					Path: entry["path"].(string),
+					URL:  entry["url"].(string),
+					Type: entry["type"].(string),
+				})
+			}
+		}
+	}
+	return files
+}
+
+func IsDependencyFile(p string) bool {
+	files := []string{
+		"package.json", "pom.xml", "build.gradle", "build.gradle.kts",
+		"go.mod", "cargo.toml", "requirements.txt", "pyproject.toml",
+		"composer.json", ".csproj", "gemfile", "gemfile.lock",
+		"yarn.lock", "pnpm-lock.yaml", "package-lock.json",
+	}
+	for _, f := range files {
+		if strings.HasSuffix(p, f) {
+			return true
+		}
+	}
+	return false
 }
